+++ Sources/StreamChatSwiftUI/ChatMessageList/Reactions/MessageActions/MessageActionsView.swift
@@ -39,7 +39,6 @@ public struct MessageActionsView: View {
                             )
                         }
                         .accessibilityLabel(action.title)
-                        .accessibilityAddTraits(.isButton)
                     } else {
                         Button {
                             if action.confirmationPopup != nil {
@@ -57,7 +56,6 @@ public struct MessageActionsView: View {
                             )
                         }
                         .accessibilityLabel(action.title)
-                        .accessibilityAddTraits(.isButton)
                     }
 
                     Divider()